// packages/database/src/mongoose/interfaces/iFileStatsCreateInput.ts
// THIS CODE WAS AUTOMATICALLY GENERATED
import {databaseTypes} from 'types';
import {Types as mongooseTypes} from 'mongoose';

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface IFileStatsCreateInput extends Omit<databaseTypes.IFileStats, '_id' | 'createdAt' | 'updatedAt'> {}